Complete A-CodeQuality
