+++ build.gradle
@@ -37,6 +37,10 @@ javafx {
     modules = [ 'javafx.controls', 'javafx.fxml']
 }
 
+run {
+    enableAssertions = true
+}
+
 application {
     // Change this to your main class.
     mainClassName = 'Launcher'
