+++ src/main/java/Launcher.java
@@ -1,3 +1,4 @@
+import duke.ui.Main;
 import javafx.application.Application;
 
 public class Launcher {
