+++ src/main/java/duke/ui/Main.java
@@ -1,3 +1,6 @@
+package duke.ui;
+
+import duke.model.Duke;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
