+++ src/main/resources/view/MainWindow.fxml
@@ -14,7 +14,7 @@
             prefWidth="541.0"
             xmlns="http://javafx.com/javafx/11.0.1"
             xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="MainWindow">
+            fx:controller="duke.ui.MainWindow">
     <children>
         <TextField fx:id="userInput"
                    layoutY="558.0"
