+++ src/test/java/duke/tasks/DeadlineTest.java
@@ -14,11 +14,11 @@ class DeadlineTest {
         // Testing string conversion for constructor 1
         FormattedDateStub date = new FormattedDateStub("s");
         Deadline d1 = new Deadline("project meeting", date);
-        assertEquals("[D][✗] project meeting (by: 20/10/3999 2345)", d1.toString());
+        assertEquals("[D][Not Done] project meeting (by: 20/10/3999 2345)", d1.toString());
 
         // Testing string conversion for constructor 2
         Deadline d2 = new Deadline("sleep", true, date);
-        assertEquals("[D][✓] sleep (by: 20/10/3999 2345)", d2.toString());
+        assertEquals("[D][Done] sleep (by: 20/10/3999 2345)", d2.toString());
     }
 
     @Test
@@ -26,7 +26,7 @@ class DeadlineTest {
         FormattedDateStub date = new FormattedDateStub("s");
         Deadline d1 = new Deadline("project meeting", date);
         d1.markAsDone();
-        assertEquals("[D][✓] project meeting (by: 20/10/3999 2345)", d1.toString());
+        assertEquals("[D][Done] project meeting (by: 20/10/3999 2345)", d1.toString());
     }
 
     @Test
