+++ src/test/java/duke/tasks/ToDoTest.java
@@ -9,17 +9,17 @@ class ToDoTest {
     @Test
     void testStringConversion() {
         // Testing string conversion for constructor 1
-        assertEquals("[T][✗] sleep", new ToDo("sleep").toString());
+        assertEquals("[T][Not Done] sleep", new ToDo("sleep").toString());
 
         // Testing string conversion for constructor 2
-        assertEquals("[T][✓] talk", new ToDo("talk", true).toString());
+        assertEquals("[T][Done] talk", new ToDo("talk", true).toString());
     }
 
     @Test
     void testIsDoneFunction() {
         ToDo t = new ToDo("sleep");
         t.markAsDone();
-        assertEquals("[T][✓] sleep", t.toString());
+        assertEquals("[T][Done] sleep", t.toString());
     }
 
     @Test
